DBZ-4339 Migrate from JUnit 4 to 5, rename utility class, fix formatting.
