+++ debezium-testing/debezium-testing-testcontainers/src/main/java/io/debezium/testing/testcontainers/util/MoreStartables.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.testing.testcontainers;
+package io.debezium.testing.testcontainers.util;
 
 import static java.util.concurrent.CompletableFuture.runAsync;
 
@@ -15,9 +15,11 @@ import org.testcontainers.lifecycle.Startable;
 import org.testcontainers.lifecycle.Startables;
 
 /**
- * Utilities for working with {@link Startable}s in a MongoDB context.
+ * Utilities for working with {@link Startable}s.
+ * <p>
+ * Extends the utilities found in {@link Startables}.
  */
-public final class MongoDbStartables {
+public final class MoreStartables {
 
     /**
      * Start every supplied {@link Startable} recursively and, to the extent possible, in parallel, waiting synchronously
@@ -25,7 +27,7 @@ public final class MongoDbStartables {
      *
      * @param startables the list of startables to {@link Startable#start())
      */
-    public static void deepStart(Stream<? extends Startable> startables) {
+    public static void deepStartSync(Stream<? extends Startable> startables) {
         try {
             Startables.deepStart(startables).get();
         }
@@ -39,8 +41,9 @@ public final class MongoDbStartables {
      * on the result.
      *
      * @param startables the list of startables to {@link Startable#stop())
+     * @see <a href="https://github.com/testcontainers/testcontainers-java/pull/1404#discussion_r1030949285">deepStop</a>
      */
-    public static void deepStop(Stream<? extends Startable> startables) {
+    public static void deepStopSync(Stream<? extends Startable> startables) {
         try {
             CompletableFuture.allOf(startables
                     .map(node -> runAsync(node::stop))
@@ -52,7 +55,7 @@ public final class MongoDbStartables {
         }
     }
 
-    private MongoDbStartables() {
+    private MoreStartables() {
         throw new AssertionError("Should not be instantiated");
     }
